wip: trying to get mAP from data
Unsuccessful, changing metric to F1 score.

mAP will be difficult to calculate for multi-class classifier (that is,
an image can have one out of the many classes; as apposed to multiple
classes per images, which is an entirely different thing).

All tools that can handle multiple classes output single precision and
recall values for the whole set of ground-true and predictions. But to
get AP, we need to calculate precision and recall on a per-class basis.